Ensure seed data is archived in Jenkins build
https://eaflood.atlassian.net/browse/WC-133

We need to be able to seed our service in none production environments, and typically we have used the test fixtures to do this.

When it comes to Jenkins though it won't see those files if we have not told it to archive them during the build. Hence this change.