Remove version from jar & simplify pom.xml (#32)
Whilst working on the systemd scripts that will be used to control the service in production, we faced an issue that we had to pass the `ExecStart` command to `/bin/bash -c` because we could not use our standard way of executing the service using `java -jar target/waste-carriers-service*.jar`.

We have to use the * because prior to this change the current version of the project would be included in the outputted jar file. Using the * meant we could use a generic command in scripts and documents and not have to worry about what the actual jar file name was.

On reflection we felt there was actually no need to record this fat in the Jar. We tag our releases in the git repo, and as the version only gets updated on major changes it was reasonably meaningless anyway.

Therefore this change updates the `pom.xml` build configuration to exclude the version info from the jar file name. This in turn means we always know what the filename will be, and we no longer have to make use of the *.

---

Also whilst we were in the `pom.xml` we reviewed what build plugins we had compared to the os-places-address-lookup project, and the current Dropwizard documentation. We realised we were depending on plugins those other projects didn't have (and to be honest we don't know why).

So we tried removing them (and updating the one we retained) and found that it still built and run fine, and in fact meant we went from a build time of 2mins 40secs to just 14secs. Result!