Update config to read from new env var when mocked (#34)
When the app is run in mock mode locally, the user facing apps and the services app can talk directly. However having deployed to our production environment and tried to get this up and running, we realised that the mocked worldpay service has to act like a public service on the internet. So it can't just tell the calling app to redirect to an internal address, it must be the public address (without port number).

Hence we now look to a new env var called `WCRS_PUBLIC_DOMAIN` and if present, we use it to formulate the url the calling app should redirect users to.