Correct expiry status job to ignore time (#38)
https://eaflood.atlassian.net/browse/WC-186

When looking into how we determine if a registration is expired for a previous story, we had to get very specific around 'when' a registration is expired. One clear thing that came out of this work was that the time a registration is made ACTIVE does not factor into it.

The problem with the `RegistrationStatusJob` as it stood prior to this change was that its filter was based on searching for any ACTIVE registration with an expiry date less than the time the job was run.

We know in production this job is run at 8pm, so what this means is if a registration was completed 3 years ago after midnight, then it gets an extra day.

So this change updates the filter to search for any ACTIVE registration with an expiry date less than tomorrows date (with time set to 00:00). So if a registration expired on the date the job is executed at 23:59:59, it would still be caught even if the job is run at 8pm.