Add test coverage plugin and integrate CodeClimate
Often it can be hard to demonstrate the technical debt a project has which may be impeding its performance. In this case we want to highlight the poor maintainability and lack of code coverage we have inherited with this project to help demonstrate why we are moving more and more functionality out of it, and how it slows us down when we changes have to be made to it.