Add location attribute to registration
https://eaflood.atlassian.net/browse/RUBY-214

As part of building the renewals service we added support for a location question. The main purpose is to

- ensure that the user is registered with the right authority
- reduce the burden on EA staff having to check for users based elsewhere who have registered with us, who they then have to confirm if this is correct.

The Waste Carriers frontend has been updated in [PR #182](https://github.com/DEFRA/waste-carriers-frontend/pull/182) to include the location question as part of the new registration journey. However as the Waste Carriers Service sits between the frontend and the database, it needs to be updated to support this new attribute.