https://eaflood.atlassian.net/browse/RUBY-817
We have now implemented mocking of Companies House in [defra-ruby-mocks](https://github.com/DEFRA/defra-ruby-mocks), a rails engine we have created to replace what is done in the service.

This is all part of the work being done to retire this app from the Waste Carriers Digital service.

This change removes the redundant mocking functionality from the project.