Remove IR Import functionality (#48)
This change removes all code related to the importing of IR registrations.

IR refers to Integrated Regulation, which was the system the Waste Carriers Registration service replaced. As part of that, we needed to import the existing IR registrations into our service.

This was a one-time operation which code was written for. That operation and task were completed years ok, so this is part of a series of PR's to remove unused functionality from the project. Removing this unused code should allow us to more easily see what the service is _actually_ still doing just prior to decommissioning it.