Remove registration status update background job (#50)
https://eaflood.atlassian.net/browse/RUBY-739

We have now implemented the expire registration job within the Rails service as part of our 'tech debt' work. This means the registration status job is no longer required and can be removed.

We have decided to remove rather than simply deactivate, so when it finally comes to archiving the project we have a clearer idea of what functions it was left performing.