Remove data export background jobs (#49)
https://eaflood.atlassian.net/browse/RUBY-737
https://eaflood.atlassian.net/browse/RUBY-738
https://eaflood.atlassian.net/browse/WC-532

We have now implemented the EPR and BOXI exports within the Rails service as part of our 'tech debt' work. This means the background jobs in this project are no longer required and can be removed. It also means the issue with memory usage caused by the project automatically gets resolved!

We have decided to remove rather than simply display, so when it finally comes to archiving the project we have a clearer idea of what functions it was left performing.